Supprime l'ancien composant IndicateurCard
